Split tests up into two modules
