+++ tests/test_anki_session.py
@@ -1,7 +1,7 @@
 # pytest-anki
 #
+# Copyright (C)  2019-2021 Aristotelis P. <https://glutanimate.com/>
 # Copyright (C)  2017-2019 Michal Krassowski <https://github.com/krassowski>
-# Copyright (C)  2019-2020 Aristotelis P. <https://glutanimate.com/>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -28,27 +28,18 @@
 #
 # Any modifications to this file must keep this entire header intact.
 
+
+"""
+Tests for the AnkiSession API
+"""
+
 import sys
 import tempfile
 from pathlib import Path
 
 import pytest
 
-from pytest_anki import AnkiSession, AnkiStateUpdate, AnkiSessionError
-
-# General tests ####
-
-
-@pytest.mark.forked
-def test_anki_session_launches(anki_session: AnkiSession):
-    from aqt import AnkiApp
-    from aqt.main import AnkiQt
-
-    assert isinstance(anki_session.app, AnkiApp)
-    assert isinstance(anki_session.mw, AnkiQt)
-    assert isinstance(anki_session.user, str)
-    assert isinstance(anki_session.base, str)
-
+from pytest_anki import AnkiSession, AnkiStateUpdate
 
 # Indirect parametrization ####
 
@@ -202,52 +193,3 @@ def test_can_preset_anki_state(anki_session: AnkiSession):
             addon.colconf_storage
             == _my_anki_state.colconf_storage[package_name]  # type: ignore
         )
-
-
-# AnkiSession API ####
-
-
-@pytest.mark.forked
-def test_collection_loading_unloading(anki_session: AnkiSession):
-    from anki.collection import Collection
-    from aqt import gui_hooks
-    
-    is_profile_loaded = False
-
-    def on_profile_did_open():
-        nonlocal is_profile_loaded
-        is_profile_loaded = True
-
-    def on_profile_will_close():
-        nonlocal is_profile_loaded
-        is_profile_loaded = False
-    
-    gui_hooks.profile_did_open.append(on_profile_did_open)
-    gui_hooks.profile_will_close.append(on_profile_will_close)
-    
-    with pytest.raises(AnkiSessionError) as exception_info:
-        _ = anki_session.collection
-    assert "Collection has not been loaded" in str(exception_info.value)
-    
-    assert is_profile_loaded is False
-    
-    with anki_session.profile_loaded():
-        assert isinstance(anki_session.collection, Collection)
-        assert is_profile_loaded is True
-
-    assert anki_session.mw.col is None
-    assert is_profile_loaded is False
-
-
-
-_deck_path = Path(__file__).parent / "samples" / "decks" / "sample_deck.apkg"
-
-
-@pytest.mark.forked
-@pytest.mark.parametrize("anki_session", [dict(load_profile=True)], indirect=True)
-def test_deck_imported(anki_session: AnkiSession):
-    collection = anki_session.collection
-    with anki_session.deck_installed(path=_deck_path) as deck_id:
-        deck = collection.decks.get(did=deck_id)
-        assert deck is not None
-        assert deck["id"] == deck_id
